add files for all card views
