+++ src/navigation/AppStack.js
@@ -11,7 +11,7 @@ import Profile from '../../src/components/Profile'
 import ProfileEdit from '../../src/components/ProfileEdit';
 import Messages from '../../src/components/Messages.js';
 import ChatMessages from '../../src/components/ChatMessages.js';
-import Cards from '../../src/components/MatchCards/index.js';
+import SwipeCards from '../components/MatchCards/SwipeCards.js';
 
 
 
@@ -37,7 +37,7 @@ export default function AppStack() {
   function MainScreen() {
     return (
       <SafeAreaView style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-        <Cards />
+        <SwipeCards />
       </SafeAreaView>
     );
   }
